/upload checked for completeness/correctness
